-Create Favourite Adapter & models -myUtils

+++ app/src/main/java/com/aquar/myaquar_egypt/Model/Login/UserInfo.java
@@ -53,7 +53,7 @@ public class UserInfo {
         this.phone = phone;
     }
 
-    public long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
